created app.ts and post test added
